Fix insidious bug in argument parsing.
This commit fixes a subtle bug in argument parsing. Previously, the
`units` paramter had `default=[1]` and `action="append"` set. This means
that every MLP run up until this fix had a topology of `1-N-1` instead
of `N-1` as expected.

To fix this bug, `default=[]` was set in the argument parser, with an
if/else block later on to allow for the default 1 node layer if the
argument was not provided.