The big config file update for the MLP script.
This commit is a massive breaking change for how the `mlp.py` script
works--changing it from doing one run and quitting, to doing entire
grid searches across several parameters and datasets. It also now is run
entirely through configuration INI files, and saves results of runs to
an SQLite3 database.